v4.12.0

+++ src_Set-OutlookSignatures/sample code/SimulateAndDeploy.ps1
@@ -1,24 +1,36 @@
 <#
-This script shows how the simulation mode of Set-OutlookSignatures can be used to deploy Outlook text signatures without client involvement.
+This sample code shows how to achieve two thing:
+	- Running simulation mode for multiple users
+  	- How to use simulation mode together with the Benefactor Circle add-on to push signatures and out-of-office replies into mailboxes,
+	  without involving end users or their devices
 
 You have to adapt it to fit your environment.
 The sample code is written in a generic way, which allows for easy adaption.
 
-Looking for support? ExplicIT Consulting (https://explicitconsulting.at) offers commercial support.
+Would you like support? ExplicIT Consulting (https://explicitconsulting.at) offers fee-based support for this and other open source code.
+
 
 Features
 	- Automate simulation mode for all given mailboxes
 	- A configurable number of Set-OutlookSignatures instances run in parallel for better performance
-	- Set default signature in Outlook Web, no matter if classic signature or roaming signatures
-	- Set internal and external out of office (OOF) message
+	- Set default signature in Outlook Web, no matter if classic signature or roaming signatures (requires the Benefacot Circle add-on)
+	- Set internal and external out of office (OOF) message (requires the Benefacot Circle add-on)
 	- Supports on-prem, hybrid and cloud-only environments
 
+
 Requirements
-	- On-prem
-		- the software needs to be run with an account that has a mailbox which is granted full access to all simulated mailboxes
+  	Follow the requirements exactly and in full. SimulateAndDeploy will not work correctly when even one requirement is not met.
+	Would you like support? ExplicIT Consulting (https://explicitconsulting.at) offers fee-based support for this and other open source code.
+
+	- For mailboxes on-prem
+		- The software needs to be run with an account that
+			- has a mailbox
+			- and is granted "full access" to all simulated mailboxes
 		- If you do not want to simulate cloud mailboxes, set $ConnectOnpremInsteadOfCloud to $true
-	- Cloud
-		- The software needs to be run with an account that has a mailbox which is granted full access to all simulated mailboxes
+	- For mailboxes in Exchange Online
+		- The software needs to be run with an account that
+			- has a mailbox
+			- and is granted "full access" to all simulated mailboxes
 		- MFA is not yet supported, but script can be adapted accordingly
 			- MFA would require interactivity, breaking the possibility for complete automation
 			- Better configure a Conditional Access Policy that only allows logon from a controlled network and does not require MFA
@@ -27,7 +39,7 @@ Requirements
 			- Option A: Create the app automatically by using the script '.\sample code\Create-EntraApp.ps1'
 				The sample code creates the app with all required settings automatically, only providing admin consent is a manual task
 			- Option B: Create the Entra app manually, with the following properties:
-				- Application permissions with admin consent
+				- Application (!) permissions with admin consent
 					- Microsoft Graph
 						- GroupMember.Read.All
 							Allows the app to list groups, read basic group properties and read membership of all groups the signed-in user has access to.
@@ -42,7 +54,7 @@ Requirements
 						- full_access_as_app
 							Allows the app to have full access via Exchange Web Services to all mailboxes without a signed-in user.
 							Required for Exchange Web Services access (read Outlook Web configuration, set classic signature and roaming signatures)
-				- Delegated permissions with admin consent
+				- Delegated (!) permissions with admin consent
 					These permissions equal those mentioned in '.\config\default graph config.ps1'
 					- Microsoft Graph
 						- email
@@ -78,11 +90,13 @@ Requirements
 	- Microsoft Word (see 'Limitations' for a scenario that does not require Word)
 	- File paths can get very long and be longer than the default OS limit. Make sure you allow long file paths.
 		- https://learn.microsoft.com/en-us/windows/win32/fileio/maximum-file-path-limitation
+	- Do not forget to adapt the "Variables" section of this script according to your needs and your configuration
+
 
 Limitations
-	- Despitze parallelization, the software runtime can be unsuited for a higher number of users. The reason usually is the Word background process.
+	- Despitze parallelization, the execution time can be too long for a higher number of users. The reason usually is the Word background process.
 		- If you use DOCX templates and/or require signatures in RTF format, Word is needed for document conversion and you can only shorten runtime by adding hardware (scale up or scale out)
-		- If you do need HTML and/or TXT signatures only, you can use the following workaround to avoid starting Word:
+		- If you do need HTML signatures only, you can use the following workaround to avoid starting Word:
 			- Use HTM templates instead of DOCX templates (parameter '-UseHtmTemplates true')
 				There are features in DOCX templates that can not be replicated HTM templates, such as applying Word specific image and text filters
 			- Do not create signatures in RTF format (parameter '-CreateRtfSignatures false')
@@ -90,6 +104,7 @@ Limitations
 		- Roaming signatures for shared mailboxes pose a general problem, as only signatures with replacement variables from the $CurrentMailbox[...]$ namespace would make sense anyhow
 #>
 
+[CmdletBinding()]
 
 # Variables
 param (
@@ -105,8 +120,14 @@ param (
 	$SetOutlookSignaturesScriptPath = '..\Set-OutlookSignatures.ps1',
 	$SetOutlookSignaturesScriptParameters = @{
 		# Do not use: SimulateUser, SimulateMailboxes, AdditionalSignaturePath, SimulateAndDeployGraphCredentialFile
-		#BenefactorCircleLicenseFile = "'\\server\share\folder\license.dll'"
-		#BenefactorCircleId = '<BenefactorCircleId>'
+		#
+		# ▼▼▼▼ The "Deploy" part of "SimulateAndDeploy" requires a Benefactor Circle license ▼▼▼▼
+		# ▼▼▼▼ Without the license, signatures can not be read from and written to mailboxes ▼▼▼▼
+		BenefactorCircleLicenseFile                   = "'\\server\share\folder\license.dll'"
+		BenefactorCircleId                            = '<BenefactorCircleId>'
+		# ▲▲▲▲ The "Deploy" part of "SimulateAndDeploy" requires a Benefactor Circle license ▲▲▲▲
+		# ▲▲▲▲ Without the license, signatures can not be read from and written to mailboxes ▲▲▲▲
+		#
 		SimulateAndDeploy                             = $false # $false simulates but does not deploy, $true simulates and deploys
 		UseHtmTemplates                               = $false
 		SignatureTemplatePath                         = "'.\sample templates\Signatures DOCX'"
@@ -120,7 +141,7 @@ param (
 		DeleteScriptCreatedSignaturesWithoutTemplate  = $true
 		SetCurrentUserOutlookWebSignature             = $true
 		SetCurrentUserOOFMessage                      = $true
-		MirrorLocalSignaturesToCloud                  = $true #Set to $false if you do not want to use this feature
+		MirrorCloudSignatures                  = $true #Set to $false if you do not want to use this feature
 		CreateRtfSignatures                           = $false
 		CreateTxtSignatures                           = $true
 		DocxHighResImageConversion                    = $true
@@ -131,7 +152,6 @@ param (
 		TrustsToCheckForGroups                        = @('*')
 		IncludeMailboxForestDomainLocalGroups         = $false
 		WordProcessPriority                           = "'Normal'"
-		Verbose                                       = '$true'
 	},
 
 	$SimulateResultPath = 'c:\test\SimulateAndDeploy',
@@ -155,6 +175,7 @@ bobby.busy@example.com;bobby.busy@example.com
 fenix.fish@example.com;fenix.fish@example.com,nat.nuts@example.com
 '@ | ConvertFrom-Csv -Delimiter ';')
 )
+
 
 # Functions
 function CreateUpdateSimulateAndDeployGraphCredentialFile {
@@ -310,11 +331,11 @@ if (-not $ConnectOnpremInsteadOfCloud) {
 	Write-Host "Connect to Graph @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
 
 	Write-Host '  Microsoft Graph'
-	$SimulateAndDeployGraphCredentialFile = Join-Path -Path $env:temp -ChildPath "$((New-Guid).guid).xml"
+	$SimulateAndDeployGraphCredentialFile = Join-Path -Path ([System.IO.Path]::GetTempPath()) -ChildPath "$((New-Guid).guid).xml"
 
 	$SetOutlookSignaturesScriptParameters['SimulateAndDeployGraphCredentialFile'] = "'$($SimulateAndDeployGraphCredentialFile)'"
 
-	Import-Module $(Join-Path -Path (Split-Path $SetOutlookSignaturesScriptPath -Parent) -ChildPath '\bin\msal.ps') -Force
+	Import-Module $(Join-Path -Path (Split-Path $SetOutlookSignaturesScriptPath -Parent) -ChildPath '\bin\MSAL.PS') -Force
 
 	$GraphConnectResult = CreateUpdateSimulateAndDeployGraphCredentialFile
 
@@ -380,31 +401,30 @@ if (-not $SimulateListCheckPositive) {
 # Overcome Word security warning when export contains embedded pictures
 # Set-OutlookSignatures handles this itself very well, but multiple instances running in the same user account may lead to problems
 # As a workaround, we define the setting before running the jobs
-Write-Host "Export Word security setting and disable it @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
-$WordRegistryVersion = [System.Version]::Parse(((((((Get-ItemProperty 'Registry::HKEY_CLASSES_ROOT\Word.Application\CurVer' -ErrorAction SilentlyContinue).'(default)' -ireplace [Regex]::Escape('Word.Application.'), '') + '.0.0.0.0')) -ireplace '^\.', '' -split '\.')[0..3] -join '.'))
-if ($WordRegistryVersion.major -eq 0) {
-	$WordRegistryVersion = $null
-} elseif ($WordRegistryVersion.major -gt 16) {
-	Write-Host "    Word version $WordRegistryVersion is newer than 16 and not yet known. Please inform your administrator. Exit." -ForegroundColor Red
-	exit 1
-} elseif ($WordRegistryVersion.major -eq 16) {
-	$WordRegistryVersion = '16.0'
-} elseif ($WordRegistryVersion.major -eq 15) {
-	$WordRegistryVersion = '15.0'
-} elseif ($WordRegistryVersion.major -eq 14) {
-	$WordRegistryVersion = '14.0'
-} elseif ($WordRegistryVersion.major -lt 14) {
-	Write-Host "    Word version $WordRegistryVersion is older than Word 2010 and not supported. Please inform your administrator. Exit." -ForegroundColor Red
-	exit 1
+if (($IsWindows -or (-not (Test-Path 'variable:IsWindows'))) -and ($SetOutlookSignaturesScriptParameters.UseHtmTemplates -inotin (1, '1', 'true', '$true', 'yes'))) {
+	Write-Host "Export Word security setting and disable it @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
+	$WordRegistryVersion = [System.Version]::Parse(((((((Get-ItemProperty 'Registry::HKEY_CLASSES_ROOT\Word.Application\CurVer' -ErrorAction SilentlyContinue).'(default)' -ireplace [Regex]::Escape('Word.Application.'), '') + '.0.0.0.0')) -ireplace '^\.', '' -split '\.')[0..3] -join '.'))
+	if ($WordRegistryVersion.major -gt 16) {
+		Write-Host "    Word version $WordRegistryVersion is newer than 16 and not yet known. Please inform your administrator. Exit." -ForegroundColor Red
+		exit 1
+	} elseif ($WordRegistryVersion.major -eq 16) {
+		$WordRegistryVersion = '16.0'
+	} elseif ($WordRegistryVersion.major -eq 15) {
+		$WordRegistryVersion = '15.0'
+	} elseif ($WordRegistryVersion.major -eq 14) {
+		$WordRegistryVersion = '14.0'
+	} elseif ($WordRegistryVersion.major -lt 14) {
+		Write-Host "    Word version $WordRegistryVersion is older than Word 2010 and not supported. Please inform your administrator. Exit." -ForegroundColor Red
+		exit 1
+	}
+
+	$WordDisableWarningOnIncludeFieldsUpdate = Get-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -ErrorAction Ignore
+
+	if (($null -eq $WordDisableWarningOnIncludeFieldsUpdate) -or ($WordDisableWarningOnIncludeFieldsUpdate -ne 1)) {
+		New-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -PropertyType DWord -Value 1 -ErrorAction Ignore | Out-Null
+		Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -Value 1 -ErrorAction Ignore | Out-Null
+	}
 }
-
-$WordDisableWarningOnIncludeFieldsUpdate = Get-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -ErrorAction Ignore
-
-if (($null -eq $WordDisableWarningOnIncludeFieldsUpdate) -or ($WordDisableWarningOnIncludeFieldsUpdate -ne 1)) {
-	New-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -PropertyType DWord -Value 1 -ErrorAction Ignore | Out-Null
-	Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -Value 1 -ErrorAction Ignore | Out-Null
-}
-
 
 # Run simulation mode for each user
 Write-Host "Run simulation mode for each user and its mailbox(es) @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
@@ -562,19 +582,20 @@ do {
 
 
 # Restore Word security setting for embedded images
-Write-Host "Restore original Word security setting @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
+if (($IsWindows -or (-not (Test-Path 'variable:IsWindows'))) -and ($SetOutlookSignaturesScriptParameters.UseHtmTemplates -inotin (1, '1', 'true', '$true', 'yes'))) {
+	Write-Host "Restore original Word security setting @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
 
-if ($null -eq $WordDisableWarningOnIncludeFieldsUpdate) {
-	Remove-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -ErrorAction Ignore
-} else {
-	Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -Value $WordDisableWarningOnIncludeFieldsUpdate -ErrorAction Ignore | Out-Null
+	if ($null -eq $WordDisableWarningOnIncludeFieldsUpdate) {
+		Remove-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -ErrorAction Ignore
+	} else {
+		Set-ItemProperty -Path "HKCU:\SOFTWARE\Microsoft\Office\$WordRegistryVersion\Word\Security" -Name DisableWarningOnIncludeFieldsUpdate -Value $WordDisableWarningOnIncludeFieldsUpdate -ErrorAction Ignore | Out-Null
+	}
 }
-
 
 Write-Host "Cleanup @$(Get-Date -Format 'yyyy-MM-ddTHH:mm:ssK')@"
 
 if (-not $ConnectOnpremInsteadOfCloud) {
-	Remove-Module msal.ps
+	Remove-Module MSAL.PS
 	Remove-Item -Force $SimulateAndDeployGraphCredentialFile -ErrorAction SilentlyContinue
 }
